refactor(console): card selector
