+++ packages/console/src/onboarding/pages/About/options.tsx
@@ -1,7 +1,4 @@
-import type {
-  CardSelectorOption,
-  MultiCardSelectorOption,
-} from '@/onboarding/components/CardSelector';
+import type { CardSelectorOption, MultiCardSelectorOption } from '@/components/CardSelector';
 
 import { CompanySize, Reason, Title } from '../../types';
